refactor: 💡 Repositoryに分離
